+++ visualization/app/codeCharta/core/data/dataService.js
@@ -9,7 +9,7 @@ import {DataModel} from "./model/dataModel";
 class DataService {
 
     /* @ngInject */
-    constructor($rootScope){
+    constructor($rootScope, deltaCalculatorService){
 
         /**
          * current data
@@ -22,6 +22,8 @@ class DataService {
          * @type {Scope}
          */
         this.$rootScope = $rootScope;
+
+        this.deltaCalculator = deltaCalculatorService;
 
     }
 
@@ -53,9 +55,8 @@ class DataService {
         //TODO unify metrics from both maps
         this.setMetrics(index);
         this.data.firstMap = this.data.revisions[index];
-        this.calculateAndStoreDeltas(this.data.firstMap, this.data.secondMap);
+        this.deltaCalculator.decorateRevisionsWithDeltas(this.data.firstMap, this.data.secondMap);
 
-        console.log("setting data", this.data);
         this.$rootScope.$broadcast("data-changed", this.data);
 
     }
@@ -63,58 +64,9 @@ class DataService {
     setSecondMapFromRevisions(index){
         this.setMetrics(index);
         this.data.secondMap = this.data.revisions[index];
-        this.calculateAndStoreDeltas(this.data.firstMap, this.data.secondMap);
-        console.log("setting data", this.data);
+        this.deltaCalculator.decorateRevisionsWithDeltas(this.data.firstMap, this.data.secondMap);
         this.$rootScope.$broadcast("data-changed", this.data);
 
-    }
-
-    //TODO speichere delta 2 minus 1 in 2 und 1 minus 2 in 1
-    calculateAndStoreDeltas(firstRevision, secondRevision) {
-
-        if(firstRevision.root && secondRevision.root) {
-            let firstRoot = d3.hierarchy(firstRevision.root);
-            let firstLeaves = firstRoot.leaves();
-            let secondRoot = d3.hierarchy(secondRevision.root);
-            let secondLeaves = secondRoot.leaves();
-
-            for (var j = 0; j < firstLeaves.length; j++) {
-                for (var k = 0; k < secondLeaves.length; k++) {
-                    if (firstLeaves[j].data.name === secondLeaves[k].data.name) {
-                        //calculate delta for those nodes attributes and push it to the second leave
-                        let firstDeltas = this.calculateAttributeListDelta(secondLeaves[k].data.attributes, firstLeaves[j].data.attributes);
-                        let secondDeltas = this.calculateAttributeListDelta(firstLeaves[j].data.attributes, secondLeaves[k].data.attributes);
-
-                        firstLeaves[j].data.deltas = firstDeltas;
-                        secondLeaves[k].data.deltas = secondDeltas;
-
-                    }
-                }
-            }
-
-        } else {
-            console.log("cant calculate deltas with 1 or less root");
-        }
-
-    }
-
-    /**
-     * Calculates deltas between two attribute lists
-     * @param {Object[]} first attributes list
-     * @param {Object[]} second attributes list
-     * @returns {Object[]} delta between first and second
-     */
-    calculateAttributeListDelta(first, second){
-        let deltas = {};
-        for (var key in second) {
-            if(key) {
-                let firstValue = first[key] ? first[key] : 0; //assume zero if no value in first
-                let secondValue = second[key];
-                let delta = secondValue - firstValue;
-                deltas[key] = delta;
-            }
-        }
-        return deltas;
     }
 
 }
